add SVGImporter, plus TPActor.autosetCenter()

+++ src/main/java/info/bschambers/toothpick/actor/TPPart.java
@@ -107,6 +107,6 @@ public abstract class TPPart implements TPEncodingHelper {
         return 0;
     }
 
-    public void translate(int x, int y) {}
+    public void translate(double x, double y) {}
 
 }
